:package: Refactor UUIDCountersMonitor
